# txscript: Shallow tx copy for signature hash calc.